allow intermediate csv columns to be reordered in the future
